requirements.txt
Django==3.1.3
django-filter==2.4.0
djangorestframework==3.12.2
Jinja2==2.11.2
django-environ==0.4.5
psycopg2==2.8.5